delete: applicationRunner
